Update Describe risks as features..md
In line 11, I changed the second user to user B... not sure if that's correct, but it seemed like it was missing something. I'm also not completely sure about how the sentence reads, but again, I'm having a little trouble correcting it because I'm not experienced with this kind of stuff.

In line 15, is the "as" that follows "typical single code" supposed to be there?

In line 17, is rebiviation a typo or a word I'm not familiar with?