Helping The Justification of Developers Focused on Quality Edits 17 July
I'm not sure what you're trying to say with this sentence: The kind of big sic of we are doing security and we want to track and we are improving security of a particular team is achievable.

Also having trouble with this paragraph: Which is a team that spends time on non functional requirements on quality stuff. And if a dev team that then ultimately they are going to have a problem. And the idea that if you know you don't have time to those things, it is totally a false.

With the final paragraph, do you mean to end saying that the investment pays off within 1-3 months, or that within 1-3 months, you would've wasted that time and money? It's communicating the same thing, but right now the first half and the second half are written in sort of opposing ways, so once you clarify, I'll be able to edit it so that the thoughts flow into one another.