Update Types of security tests..md
Hi Dinis,

I want to check a couple of things on this one with you:

1. Do you prefer NetSec to network security, to have consistency with AppSec?
2. Can you confirm that I have understood your meaning correctly in the way I've edited the second paragraph, i.e. "The pen test might also have been done by somebody who most likely cannot really code very well, but might have a very good security instinct or security knowledge from the point of view of network security."

I am unable to add labels to issues for my questions, please let me know which is easier for you, for me to add them here, or to continue to add them as issues, unlabelled. Thanks.